common: Move time query to common header.
